+++ apps/catalog/lib/use_case/cores/designsystem/theme/containment/alert_dialog.dart
@@ -3,10 +3,13 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart';
 
+/// AlertDialog
+/// https://api.flutter.dev/flutter/material/AlertDialog-class.html
+
 @UseCase(
   name: 'Default',
   type: AlertDialog,
-  path: '[Theme Preview]/Dialog',
+  path: '[Theme Preview]/Containment',
 )
 Widget alertDialog(BuildContext context) => Scaffold(
       body: Center(
@@ -52,7 +55,7 @@ Widget alertDialog(BuildContext context) => Scaffold(
 @UseCase(
   name: 'Adaptive',
   type: AlertDialog,
-  path: '[Theme Preview]/Dialog',
+  path: '[Theme Preview]/Containment',
 )
 Widget adaptiveAlertDialog(BuildContext context) => Scaffold(
       body: Center(
@@ -90,88 +93,3 @@ Widget adaptiveAlertDialog(BuildContext context) => Scaffold(
         ),
       ),
     );
-
-@UseCase(
-  name: 'DatePicker',
-  type: DatePickerDialog,
-  path: '[Theme Preview]/Dialog',
-)
-Widget datePicker(BuildContext context) {
-  return Scaffold(
-    body: Center(
-      child: ElevatedButton(
-        onPressed: () async {
-          final first = DateTime.now();
-          final last = first.add(const Duration(days: 6));
-          final result = await showDatePicker(
-            context: context,
-            firstDate: first,
-            lastDate: last,
-          );
-          if (context.mounted) {
-            ScaffoldMessenger.maybeOf(context)
-                ?.showSnackBar(SnackBar(content: Text('Result: $result')));
-          }
-        },
-        child: const Text('Show DatePicker'),
-      ),
-    ),
-  );
-}
-
-@UseCase(
-  name: 'TimePicker',
-  type: TimePickerDialog,
-  path: '[Theme Preview]/Dialog',
-)
-Widget timePicker(BuildContext context) {
-  return Scaffold(
-    body: Center(
-      child: ElevatedButton(
-        onPressed: () async {
-          final result = await showTimePicker(
-            context: context,
-            initialTime: TimeOfDay.now(),
-          );
-          if (context.mounted) {
-            ScaffoldMessenger.maybeOf(context)?.showSnackBar(
-              SnackBar(content: Text('Result: ${result?.format(context)}')),
-            );
-          }
-        },
-        child: const Text('Show TimePicker'),
-      ),
-    ),
-  );
-}
-
-@UseCase(
-  name: 'DateRangePicker',
-  type: DateRangePickerDialog,
-  path: '[Theme Preview]/Dialog',
-)
-Widget dateRangePicker(BuildContext context) {
-  return Scaffold(
-    body: Center(
-      child: ElevatedButton(
-        onPressed: () async {
-          final first = DateTime.now();
-          final last = first.add(const Duration(days: 6));
-          final result = await showDateRangePicker(
-            context: context,
-            firstDate: first,
-            lastDate: last,
-          );
-          if (context.mounted) {
-            ScaffoldMessenger.maybeOf(context)?.showSnackBar(
-              SnackBar(
-                content: Text('Result: ${result?.start} - ${result?.end}'),
-              ),
-            );
-          }
-        },
-        child: const Text('Show DateRangePicker'),
-      ),
-    ),
-  );
-}
