+++ apps/catalog/lib/use_case/cores/designsystem/theme/navigation/bottom_app_bar.dart
@@ -2,36 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart';
 
-@UseCase(
-  name: 'Default',
-  type: AppBar,
-  path: '[Theme Preview]/AppBar',
-)
-Widget appBar(BuildContext context) => Scaffold(
-      appBar: AppBar(
-        title: const Text('AppBar Sample'),
-        actions: <Widget>[
-          IconButton(
-            icon: const Icon(Icons.add_alert),
-            tooltip: 'Show Snackbar',
-            onPressed: () {
-              ScaffoldMessenger.of(context).showSnackBar(
-                const SnackBar(content: Text('This is a snackbar')),
-              );
-            },
-          ),
-        ],
-      ),
-      body: ListView.builder(
-        itemCount: 100,
-        itemBuilder: (context, index) => ListTile(
-          title: const Text('Title'),
-          subtitle: const Text('Subtitle'),
-          onTap: () {},
-        ),
-      ),
-    );
-
 class _BottomAppBarDemo extends StatefulWidget {
   const _BottomAppBarDemo();
 
@@ -220,9 +190,12 @@ class _DemoBottomAppBar extends StatelessWidget {
   }
 }
 
+/// BottomAppBar
+/// https://api.flutter.dev/flutter/material/BottomAppBar-class.html
+
 @UseCase(
   name: 'BottomAppBar',
   type: BottomAppBar,
-  path: '[Theme Preview]/AppBar',
+  path: '[Theme Preview]/Navigation',
 )
 Widget bottomAppBar(BuildContext context) => const _BottomAppBarDemo();
