+++ apps/catalog/lib/use_case/cores/designsystem/theme/selection/chip.dart
@@ -4,7 +4,7 @@ import 'package:widgetbook_annotation/widgetbook_annotation.dart';
 @UseCase(
   name: 'Chip',
   type: Chip,
-  path: '[Theme Preview]',
+  path: '[Theme Preview]/Selection',
 )
 Widget chip(BuildContext context) => const Chip(
       avatar: CircleAvatar(),
